Devide Contact container into web and native, migrate Contact to TS. Move mutations out the Contact container
