+++ packages/client/src/modules/contact/types/index.ts
@@ -1,0 +1,15 @@
+/* Entities */
+interface Contact {
+  name: string;
+  email: string;
+  content: string;
+}
+
+/* ComponentProps */
+interface ContactOperations {
+  contact: (contactData: Contact) => boolean;
+}
+
+interface ContactProps extends ContactOperations {}
+
+export { Contact, ContactOperations, ContactProps };
